refactor: create separate file for rag cli display helpers
