+++ src/mmore/cli.py
@@ -1,6 +1,6 @@
 import os
 
-# To have the fork-handler messages ("FD from fork parent still in poll list")
+# To not have the fork-handler messages ("FD from fork parent still in poll list")
 os.environ.setdefault("GRPC_VERBOSITY", "ERROR")
 
 from typing import Optional
